[refactor] test 수정
